+++ pymc/tests/distributions/test_timeseries.py
@@ -43,8 +43,7 @@ from pymc.distributions.timeseries import (
 )
 from pymc.model import Model
 from pymc.sampling import sample
-from pymc.sampling_forward import sample_posterior_predictive
-from pymc.sampling_utils import draw
+from pymc.sampling_forward import draw, sample_posterior_predictive
 from pymc.tests.distributions.util import assert_moment_is_expected
 from pymc.tests.helpers import select_by_precision
